+++ src/Http/Router.php
@@ -226,6 +226,8 @@ class Router
                 }
             }
         }
+
+        return [];
     }
 
     /**